+++ post-install.d/git.sh
@@ -87,10 +87,10 @@ printf "Configuring pull behaviour...\n"
 git config --global pull.ff only
 printf "Setting up some aliases...\n"
 git config --global alias.unstage 'reset -q HEAD --'
-git config --global alias.flog "log --all --graph --oneline --format=format:'%C(bold yellow)%h%C(r) %an: %C(bold)%s%C(r) %C(auto)%d%C(r)'"
-git config --global alias.eflog "log --all --graph --oneline --format=format:'%C(bold yellow)%h%C(r) %ae: %C(bold)%s%C(r) %C(auto)%d%C(r)'"
+git config --global alias.flog "log --all --graph --oneline --format=format:'%C(bold yellow)%h%C(auto) %an: %C(bold)%s%C(auto) %C(auto)%d'"
+git config --global alias.eflog "log --all --graph --oneline --format=format:'%C(bold yellow)%h%C(auto) %ae: %C(bold)%s%C(auto) %C(auto)%d'"
 git config --global alias.slog 'log --show-signature'
-git config --global alias.sflog "log --all --graph --oneline --format=format:'%C(bold yellow)%h%C(r) %C(bold green)%G?%C(r) %an: %C(bold)%s%C(r) %C(auto)%d%C(r)'"
+git config --global alias.sflog "log --all --graph --oneline --format=format:'%C(bold yellow)%h%C(auto) %C(bold green)%G?%C(auto) %an: %C(bold)%s%C(r) %C(auto)%d'"
 git config --global alias.mkst 'stash push -u'
 git config --global alias.popst 'stash pop "stash@{0}" -q'
-git config --global alias.list-ignored 'ls-files -i --exclude-standard'
+git config --global alias.list-ignored 'ls-files -i -o --exclude-standard'