+++ post-install.d/zsh.sh
@@ -5,8 +5,9 @@ Separate
 printf "Successfully installed \e[36mzsh\e[00m, configuring...\n"
 
 # Create .zshrc files
-     [ ! -f ~/.zshrc     ] && cat "$script_location/samples/zshrc" | tee ~/.zshrc ~/.zshrc-og >/dev/null
-sudo [ ! -f /root/.zshrc ] && cat "$script_location/samples/zshrc" | sudo tee /root/.zshrc /root/.zshrc-og >/dev/null
+     [ ! -f ~/.zshrc         ] && cat "$script_location/samples/zshrc" | tee ~/.zshrc ~/.zshrc-og >/dev/null
+sudo [ ! -f /root/.zshrc     ] && cat "$script_location/samples/zshrc" | sudo tee /root/.zshrc /root/.zshrc-og >/dev/null
+sudo [ ! -f /etc/skel/.zshrc ] && cat "$script_location/samples/zshrc" | sudo tee /etc/skel/.zshrc >/dev/null
 
 # Offer to install powerline-shell
 read -rp "$(printf "Do you want to install \e[01mPowerline Shell\e[00m? (y/N) ")"
@@ -70,4 +71,6 @@ read -rp "Do you want to make `tput setaf 6`Z-Shell`tput sgr0` your default shel
 if [[ ${REPLY,,} == "y" ]] || [ -z $REPLY ]; then
 	sudo usermod -s $(which zsh) $USER
 	sudo usermod -s $(which zsh) root
+
+	sudo sed -i "s|SHELL=.*|SHELL=$(which zsh)|g" /etc/default/useradd
 fi